adding style for footer
